Ensure that target files are larger than 500MB; Set the splits sizes
Before target file was processed once over long periods of time, but
this is somewhat unefficient because the log can grow ridiculously big.
Instead let's run the script over short periods checking each time the
size of the file and splitting if necessary.

Splits sizes were defined in lines instead of bytes.

+++ autosplit.sh
@@ -1,26 +1,36 @@
 #! /bin/bash
 
 # Description:
-# Split a file (usually a log) into parts and compress them. 
-# It's intended to be executed over long periods as a scheduled
-# cron job
+# Split a file (usually a log) larger than 500MB into parts and compress them. 
+# It's intended to be executed as a scheduled cron job
 #
 # Args:
 # $1 -> path/to/file
 
+timestamp=`date`
+
+# First check that file exists and it's no 0 sized
 if [ ! -s $1 ]; then 
-	timestamp=`date`
 	echo "${timestamp} - No file was especified or its size is 0." >> autosplit.log
 	exit
 fi
+
+# Just operate on files larger than 500MB
+minimumsize=500000  # 500MB
+actualsize=$(du -k "$1" | cut -f 1)
+if [ $actualsize -le $minimumsize ]; then
+	echo "${timestamp} - The file was not big enough." >> autosplit.log
+    exit
+fi
+
 
 # Remove older (should be really old) splits
 if [[ (`ls | grep .tar.gz`) ]]; then
 	rm -f split.log0*.tar.gz
 fi
 
-# Split the file and compress parts
-split -d -l 300 $1 split.log  
+# Split the file in 100MB parts and compress
+split -d -b 100000000 $1 split.log  
 
 for entry in *.log0*; do
 	tar cfz "${entry}.tar.gz" "$entry"
